Tests/LibWeb/Text/expected/css/computed-values-edge-cases.txt

overflow-clip-margin: content-box 15px
degenerate auto aspect-ratio height: 50
